requirements.txt
streamlit
pyautogui
keyboard
python-dotenv
openai~=0.28
pypdf